review works, administration works, pretty good!

+++ app/controllers/TagController.php
@@ -1,35 +1,43 @@
 <?php
 
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+
 class TagController extends BaseController {
 
 	/**
 	 * Displays the tags for admin.	 
-	 * @param $userId the user id of the admin
 	 */
-	public function showTagsListPage() 
-	{
-		return View::make('user.admin.tags')->with('tags', Tag::take(100)->get());
+	public function showTagsListPage() {
+		return View::make('user.admin.tags.index')->with('tags', Tag::take(100)->get());
 	}
 
 	/**
 	 * Displays a page to add tag for admin.
 	 */
-	public function showAddTagPage() 
-	{
-		return View::make('user.admin.tags-new');
+	public function showAddTagPage() {
+		return View::make('user.admin.tags.new');
 	}
 
 	/**
 	 * Handles a post request of add tag.
 	 */
-	public function addTag() 
-	{
-		// TODO Better
-		$name = Input::get('name');
+	public function addTag() {
+		$name = Input::get('name', '');
+		if (!(strlen($name) > 0))
+			return Redirect::back()
+				->withInput()
+				->with('error', 'Du måste ange taggens namn.');
+
+		if (Tag::where('name', '=', $name)->count() > 0)
+			return Redirect::route('admin-tags')
+				->withInput()
+				->with('error', 'Taggen fanns redan och lades därför inte till.');
+
 		$tag = new Tag;
 		$tag->name = $name;
-		$tag->token = $this->generateToken($name);
+		$tag->token = $this->generateToken($name); // Move generateToken to separate class
 		$tag->save();
+
 		return Redirect::route('admin-tags')->with('success', 'Taggen skapades.'); // TODO Show
 	}
 
@@ -38,8 +46,14 @@ class TagController extends BaseController {
 	 */
 	public function showDeleteTagPage($token) 
 	{
-		$tag = Tag::where('token', '=', $token)->firstOrFail(); // TODO Make sure no fail...
-		return View::make('user.admin.tags-delete')->with('tag', $tag);
+		try {
+			$tag = Tag::where('token', '=', $token)->firstOrFail();
+		} catch(ModelNotFoundException $e) {
+		    return Redirect::route('admin-tags')
+		    	->with('error', 'Taggen som skulle tas bort hittades inte.');
+		}
+		
+		return View::make('user.admin.tags.delete')->with('tag', $tag);
 	}
 
 	/**
@@ -47,10 +61,13 @@ class TagController extends BaseController {
 	 */
 	public function deleteTag() 
 	{
-		// TODO Better
+		// Check which button was pressed, only 'yes' should continue
+		if (!Input::get('yes'))
+			return Redirect::route('admin-tags')->with('warning', 'Taggen togs inte bort.');
+
 		$token = Input::get('tag-token');
 		Tag::where('token', '=', $token)->delete();
-		return Redirect::route('admin-tags')->with('success', 'Taggen togs bort.'); // TODO Show
+		return Redirect::route('admin-tags')->with('success', 'Taggen togs bort.');
 	}
 
 	/**
@@ -58,8 +75,13 @@ class TagController extends BaseController {
 	 */
 	public function showEditTagPage($token) 
 	{
-		$tag = Tag::where('token', '=', $token)->firstOrFail(); // TODO Make sure no fail...
-		return View::make('user.admin.tags-edit')->with('tag', $tag)->with('success', 'Taggen togs bort.');
+		try {
+			$tag = Tag::where('token', '=', $token)->firstOrFail();
+		} catch(ModelNotFoundException $e) {
+		    return Redirect::route('admin-tags')
+		    	->with('error', 'Taggen som skulle ändras hittades inte.');
+		}
+		return View::make('user.admin.tags.edit')->with('tag', $tag);
 	}
 
 	/**
@@ -67,18 +89,46 @@ class TagController extends BaseController {
 	 */
 	public function editTag() 
 	{
-		// TODO Better
 		$token = Input::get('token');
-		$tag = Tag::where('token', '=', $token)->firstOrFail();
+		try {
+			$tag = Tag::where('token', '=', $token)->firstOrFail();
+		} catch(ModelNotFoundException $e) {
+		    return Redirect::back()
+		    	->withInput()
+		    	->with('error', 'Taggen som skulle ändras hittades inte.');
+		}
+
+		if (!(strlen(Input::get('name', '')) > 0))
+			return Redirect::back()
+				->withInput()
+				->with('error', 'Du måste ange taggens namn.');
+
 		$tag->name = Input::get('name');
+		$tag->token = $this->generateToken($name); // TODO Move function to own class
 		$tag->save();
-		return Redirect::route('admin-tags')->with('success', 'Taggen uppdaterades.'); // TODO Show
+
+		return Redirect::route('admin-tags')->with('success', 'Taggen uppdaterades.');
+	}
+
+	/**
+	 * Displays a list of all the PM for tag.
+	 * @param $tag the tag
+	 * @param $page the page, 1 by deafult
+	 */
+	public function showTagPMListPage($tag, $page = 1)
+	{
+		return View::make('tag.show')
+			->with('tag', $tag)
+			->with('page', $page);
 	}
 
 	/**
      * Generates a valid token.
+     * @param name the name of the original name
+     * @param delimiter the character to replace bad chars with, default is '-'
 	 */
 	private function generateToken($name, $delimiter = '-') {
+		// TODO Move
 		setlocale(LC_ALL, 'en_US.UTF8');
 		$clean = iconv('UTF-8', 'ASCII//TRANSLIT', $name);
 		$clean = preg_replace("/[^a-zA-Z0-9\/_|+ -]/", '', $clean);
@@ -93,14 +143,19 @@ class TagController extends BaseController {
 		return $clean;
 	}
 
-
 	/**
-	 * Displays a list of all the PM for tag.
-	 * @param $tag the tag
-	 * @param $page the page, 1 by deafult
+	 * Returns list of tags matching the query in json.
 	 */
-	public function showTagPMListPage($tag, $page = 1)
-	{
-		return View::make('tag.show')->with('tag', $tag)->with('page', $page);
+	public function tagsAutocomplete() {
+		$searchQuery = Input::get('q');
+		$tags = Tag::where('name', 'LIKE', '%' . $searchQuery . '%')->take(7)->get();
+		$result = array();
+		foreach($tags as $tag) {
+			$obj = new stdClass();
+			$obj->id = $tag->id;
+			$obj->name = $tag->name;
+			$result[] = $obj;
+		}
+		return json_encode($result);
 	}
 }
